finished Rest Ai application

+++ controllers/feed.js
@@ -1,5 +1,7 @@
 const { validationResult } = require('express-validator')
+
 const Post = require('../models/post')
+const User = require('../models/user')
 const fs = require('fs')
 const path = require('path')
 exports.getPosts = (req, res, next) => {
@@ -37,17 +39,25 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/')
   const title = req.body.title
   const content = req.body.content
+  let creator
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Samuel' },
+    creator: req.userId
   })
 
   post.save().then(result => {
+    return User.findById(req.userId)  
+  }).then(user => {
+    creator = user
+    user.posts.push(post)
+    return user.save()
+  }).then(result => {
     res.status(201).json({
       message: 'Post created succesfuly',
-      post: result
+      post: post,
+      creator: { _id: creator._id, name: creator.name }
     })
   }).catch(err => {
     if (!err.statusCode) {
@@ -100,6 +110,11 @@ exports.updatePost = (req, res, next) => {
       err.statusCode = 422
       throw err
     }
+    if (post.creator.toString() !== req.userId) {
+      const err = new Error('Not authorized')
+      err.statusCode = 403
+      throw err
+    }
     if (imageUrl !== post.imageUrl) {
       clearImage(post.imageUrl)
     }
@@ -125,12 +140,19 @@ exports.deletePost = (req, res, next ) => {
       err.statusCode = 422
       throw err
     }
-    // check logged in user
-
+    if (post.creator.toString() !== req.userId) {
+      const err = new Error('Not authorized')
+      err.statusCode = 403
+      throw err
+    }
     clearImage(post.imageUrl)
     return Post.findByIdAndRemove(postId)
   }).then(result => {
-    console.log(result)
+    return User.findById(req.userId)
+  }).then(user => {
+    user.posts.pull(postId)
+    return user.save()
+  }).then(result => {
     res.status(200).json({ message: 'Deleted Post!' })
   }).catch(err => {
     if (!err.statusCode) {
@@ -140,6 +162,54 @@ exports.deletePost = (req, res, next ) => {
   })
 }
 
+exports.getStatus = (req, res, next ) => {
+  if (!req.userId) {
+    const error = new Error('No user')
+    error.statusCode = 404
+    throw error
+  }
+  User.findById(req.userId).then(user => {
+    if (!user) {
+      const error = new Error('No user')
+      error.statusCode = 404
+      throw error
+    }
+    return user.status
+  }).then(status => {
+    res.status(200).json({ status: status })
+  }).catch(err => {
+    if (!err.statusCode) {
+      err.statusCode = 500
+    }
+    next(err)
+  })
+}
+
+exports.postStatus = (req, res, next ) => {
+  const updatedStatus = req.body.status
+  if (!req.userId) {
+    const error = new Error('No user')
+    error.statusCode = 404
+    throw error
+  }
+  User.findById(req.userId).then(user => {
+    if (!user) {
+      const error = new Error('No user')
+      error.statusCode = 404
+      throw error
+    }
+
+    user.status = updatedStatus
+    return user.save()
+  }).then(result => {
+    res.status(200).json({ message: 'Updated status', status: result.status })
+  }).catch(err => {
+    if (!err.statusCode) {
+      err.statusCode = 500
+    }
+    next(err)
+  })
+}
 const clearImage = (filePath) => {
   console.log(filePath)
   filePath = path.join(__dirname, '..', filePath)
